v4
POST AND LIKE ARE CREATED BUT STILL HAVE PROBLEM WHIT FETCHING listof likes and comment for one post

+++ Backend/src/main/java/Formcom/example/Taaw3iya/dao/entities/User.java
@@ -2,7 +2,6 @@ package Formcom.example.Taaw3iya.dao.entities;
 
 
 import jakarta.persistence.*;
-import jakarta.persistence.criteria.CriteriaBuilder;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -17,7 +16,7 @@ import Formcom.example.Taaw3iya.dao.enums.Role;
 
 import java.util.Collection;
 import java.util.Date;
-import java.util.List;
+
 @Setter
 @Getter
 @Table(name = "users")
@@ -53,8 +52,8 @@ public class User implements UserDetails{
     public Collection<? extends GrantedAuthority> getAuthorities() {
         return role.getAuthorities();
     }
-    public Integer getIDD(){
-        return this.id;
+    public Long getIDD(){
+        return Long.valueOf(this.id);
     }
 
     // @Override
